# src/main/java/com/guvi/inventory_order_mgt/controller/AuthController.java
package com.guvi.inventory_order_mgt.controller;

import com.guvi.inventory_order_mgt.dto.AuthResponse;
import com.guvi.inventory_order_mgt.dto.LoginRequest;
import com.guvi.inventory_order_mgt.dto.RegisterRequest;
import com.guvi.inventory_order_mgt.service.AuthService;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

/*
Endpoints
`POST /api/auth/register`, `POST /api/auth/login`
*/
@RestController
@RequestMapping("/api/auth")
public class AuthController {

    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @PostMapping("/register")
    public ResponseEntity<AuthResponse> register(@Valid @RequestBody RegisterRequest request) {
        return ResponseEntity.status(HttpStatus.CREATED).body(authService.register(request));
    }

    @PostMapping("/login")
    public ResponseEntity<AuthResponse> login(@Valid @RequestBody LoginRequest request) {
        return ResponseEntity.ok(authService.login(request));
    }
}